Add pages with partial translation to region condition
