+++ integreat_cms/cms/templates/region_condition/region_condition.html
@@ -25,16 +25,19 @@
                         {% translate "Region" %}
                     </th>
                     <th>
-                        {% translate "Number of pages" %}
-                    </th>
-                    <th>
-                        {% translate "Number of outdated pages" %}
+                        {% translate "Number of broken links" %}
                     </th>
                     <th>
                         {% translate "Number of pages under HIX threshold" %}
                     </th>
                     <th>
-                        {% translate "Number of broken links" %}
+                        {% translate "Number of pages" %}
+                    </th>
+                    <th>
+                        {% translate "Number of pages with at least one missing or outdated translation" %}
+                    </th>
+                    <th>
+                        {% translate "Number of outdated pages" %}
                     </th>
                     <th>
                         {% translate "Number of active languages besides default language" %}
