+++ integreat_cms/locale/de/LC_MESSAGES/django.po
@@ -8305,13 +8305,8 @@ msgstr "CSV herunterladen"
 
 #: cms/templates/region_condition/region_condition.html
 #: cms/views/region_condition/region_condition_actions.py
-msgid "Number of pages"
-msgstr "Anzahl an Seiten"
-
-#: cms/templates/region_condition/region_condition.html
-#: cms/views/region_condition/region_condition_actions.py
-msgid "Number of outdated pages"
-msgstr "Anzahl veralteter Seiten"
+msgid "Number of broken links"
+msgstr "Anzahl fehlerhafter Links"
 
 #: cms/templates/region_condition/region_condition.html
 msgid "Number of pages under HIX threshold"
@@ -8319,8 +8314,19 @@ msgstr "Anzahl an Seiten unter HIX Schwellenwert"
 
 #: cms/templates/region_condition/region_condition.html
 #: cms/views/region_condition/region_condition_actions.py
-msgid "Number of broken links"
-msgstr "Anzahl fehlerhafter Links"
+msgid "Number of pages"
+msgstr "Anzahl an Seiten"
+
+#: cms/templates/region_condition/region_condition.html
+#: cms/views/region_condition/region_condition_actions.py
+msgid "Number of pages with at least one missing or outdated translation"
+msgstr ""
+"Anzahl Seiten mit mindestens einer fehlenden oder veralteten Übersetzung"
+
+#: cms/templates/region_condition/region_condition.html
+#: cms/views/region_condition/region_condition_actions.py
+msgid "Number of outdated pages"
+msgstr "Anzahl veralteter Seiten"
 
 #: cms/templates/region_condition/region_condition.html
 #: cms/views/region_condition/region_condition_actions.py
@@ -11070,11 +11076,6 @@ msgid "Number of low hix pages"
 msgstr "Anzahl Seiten mit niedrigem HIX-Wert"
 
 #: cms/views/region_condition/region_condition_actions.py
-msgid "Number of pages with at least one missing or outdated translation"
-msgstr ""
-"Anzahl Seiten mit mindestens einer fehlenden oder veralteten Übersetzung"
-
-#: cms/views/region_condition/region_condition_actions.py
 msgid "Has translation package been booked?"
 msgstr "Wurde das Add-On für Übersetzungen gebucht?"
 
@@ -11590,6 +11591,10 @@ msgid "This page belongs to another region ({})."
 msgstr ""
 "Diese Seite konnte nicht importiert werden, da sie zu einer anderen Region "
 "gehört ({})."
+
+#~ msgid "Number of pages with at least one outdated translation"
+#~ msgstr ""
+#~ "Anzahl Seiten mit mindestens einer fehlenden oder veralteten Übersetzung"
 
 #~ msgid "attachment map"
 #~ msgstr "Anhangs-Mapping"
